apps/ml_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.6
statsmodels>=0.14.0
lifelines>=0.27.8
shap>=0.45.0
joblib>=1.3.0
asyncpg>=0.29.0
apscheduler>=3.10.4,<3.11
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.2
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
httpx>=0.27.0
